feat: Add stylistic features
Added colors for background and text. In order to adjust the text of the
header to correspond with the width of the action element it was
necessary to address the .navbar, .heading and .action elements
all together instead of putting them in an additional container because
I wasn't able to find a way that wouldn't effect the width of the .quote
element.

Up to this point the page isn't really responsive.